Commented a bunch of stuff in LoadHardwareClass & Teleop_Main_, and moved all previous experimentation into it's own package.
